constants
*-> Reading the headers into the prototype.
*-> Reading most constants except for KNs and the debug info.
!-> some refactoring needs to happen on ArrayPart, HashPart.
!-> constants_table in prototype needs its own file.
!-> lua_table probably needs a format function.
!!-> prototyper needs tested.  byte reading for a mock luajit compiled file needs tested. constant section reading needs tested. instruction reading needs tested.

+++ src/bytecode_instruction.rs
@@ -20,14 +20,14 @@ impl Registers {
 }
 
 pub struct BytecodeInstruction {
-    op: u32,
+    op: u8,
     registers: Registers,
 }
 
 impl BytecodeInstruction {
     pub const INSTRUCTION_SIZE: u8 = 4;
 
-    pub fn new(op: u32, a: u8, c: u8, b: u8) -> BytecodeInstruction {
+    pub fn new(op: u8, a: u8, c: u8, b: u8) -> BytecodeInstruction {
         BytecodeInstruction {
             op: op,
             registers: Registers::new(a, c, b),
